Use @FreeBuilder to generate the Builder API for the Metadata and Metadata.Property types.
To do this, the two classes were temporarily annotated with @FreeBuilder, and the generated classes checked in to the codebase as-is. These will need to be deleted and recreated if the APIs change. This ensures a higher-quality API with less potential for bugs, at the expense of a non-obvious manual procedure.

This closes issue #27.