Support generic features in SourceBuilder
SourceBuilder can now be fluently extended with arbitrary features. When queried for a feature type, it will be determined based on the processing environment. Unit tests can instead provide a fixed feature set; SourceBuilder will then return either the exact feature it was constructed with, or the fallback for that feature type. This should mean less code churn when new features need to be extracted from the environment, such as new libraries.

SourceLevel and Guava support are now implemented as features rather than being hard-coded into SourceBuilder.