Add check, mapper and mutator methods
Currently added for properties of type:
 * list
 * set
 * map
 * optional
 * default

The new mutateFoo methods require deprecating the old approach of overriding a mutator method (addFoo, putFoo) to add constraint checking for collections, since the check now needs to be applied for mutation operations that cannot be built on add/put. Issue a warning (and disable the mutator method) if the old approach is used, for backwards compatibility.