Don't throw IllegalStateExceptions from put/remove
FreeBuilder copied ImmutableMap.Builder's lead and threw IllegalStateExceptions when putting an entry for a key already present in the map. As FreeBuilder types are frequently subject to modifications rather than always created from scratch, this has turned out to be a bad match for real-world usage. (Additionally, it blocks implementation of issue #6.) Since the behaviour has changed to be more permissive, i.e. to allow actions that would previously have thrown exceptions, this change is backwards-compatible.