Allow builder classes to be declared as abstract
It is useful if Builder classes can be declared as abstract, for then
(for example) generic types can be discovered at runtime.

This commit ensures that no builder factory is registered for such a
class. Current behaviour is to generate invalid Java, since such a class
cannot be instantiated with its no-args constructor.