Use source level in FunctionPackage
Eclipse is bugged, and will never return source level > 7. Previously we hacked around this by only checking for the presence of a java.util.function type in the processing environment, but this means we generate Java 8 code in javac 8 even when source level is set to 6/7 unless users set the bootstrap classpath correctly. Now we check for the presence of the Eclipse round environment on the stack trace, and only skip the source level check if we find it.

This fixes #148.