Reduce ImmutableList duplication
When Guava is available, list properties are now initially ImmutableList instances in builders, optimistically avoiding unnecessary array allocation and ImmutableList duplication. If an object is cloned and modified using `Builder.from` without changing the list property, no ArrayList will be allocated: a single ImmutableList instance will be used throughout.

Once an ArrayList has been allocated, it will be used until the builder is discarded. While it might potentially be a memory saving to replace it with an ImmutableList once `build()` is called, in case multiple instances are minted without further modification, it would be difficult to retain the "live view" contract of the builder's getter method, which is currently done by returning an unmodifiable wrapper around the ArrayList.