Support non-bean naming convention
Allow the user to drop the 'get'/'is' prefixes from their value type declaration. When using this "prefixless" convention, we drop the "get" and "set" prefixes from the generated Builder API to match. The existing behaviour is now referred to as "bean convention".

Code to detect and understand the naming convention is in the `o.i.f.processor.naming` package. Most code generation changes are localised to the BuilderMethods utility class we factored out for this purpose. To keep Jackson usage simple, we now automatically annotate getter methods on the value type; previously, Jackson was happily spotting the 'get'/'is' prefixes itself.

New tests are all copies of existing tests modified to use prefixless convention