Add SourceLevel.JAVA_8
Previously, the FunctionPackage feature was standing in for this constant. After this refactoring, FunctionPackage defers to SourceLevel's logic to decide whether it's available or not, and tests can use simply `SourceLevel.JAVA_8` instead of both `SourceLevel.JAVA_7` and `FunctionPackage.AVAILABLE`. `FunctionPackage.lambdasAvailable()` has moved to the more obvious `SourceLevel.hasLambdas()`.

To allow FunctionPackage to determine whether SourceLevel is available, the FeatureType abstract class now takes a FeatureSet argument on its methods. Individual features are responsible for not creating cyclic dependencies. JavaxPackage uses this to disable itself by default in Java 8 tests, though it continues to defer to the processing environment in production.