Add support for SortedSet properties
These are treated very similarly to regular Set properties, except a protected `setComparatorForX` method is generated to allow a non-default comparator to be set at construction.

You might expect NavigableSet to be supported as well, since TreeSet and ImmutableSortedSet implement both, but unfortunately Collections only gained an `unmodifiableNavigableSet` method in Java 8. Support on Java 8+ may come in a later PR.