Cast Collection to List when necessary
ListPropertyFactory.addMergeFromValue was assigning the result of the value type's getter method directly to the builder's field, which fails to work if the type is a Collection, or has a wildcard. Instead, cast it to a List with ImmutableList.copyOf (since this code path is only incurred when the value has been confirmed to be an ImmutableList). SetPropertyFactory.addMergeFromValue had the same issue with wildcards, with the same fix.