Fix a bug detecting overridden methods in generic buildable types
Even though the type variable in a value type and its builder may look identical, javac doesn't consider them as the same, so the code checking to see if a method was overridden (for instance, to decide whether the call to addX can be inlined in the list mutate method) was failing to spot overrides in generic buildable types. Fix this by erasing the parameter types before passing them to isSameType.