Fix variable naming conflicts in toString code
When constructing the toString representation of two or more properties, one of them optional or nullable, without Guava available, the variable names "result" and "separator" were used raw in the code. If there was a property with either name, this would result in the buffer being pasted back into itself, rather than using the property value. The same bug also affects the toString method of partials. Fix is to use the new variable declaration framework.