Stop generating temporary arrays and strings in toString methods
Rationalise the various toString-generating code paths into a single method that uses one of three strategies:

 1) Simple concatenation, if all properties are always output
 2) A single StringBuilder, if at least one property is always output
 3) A StringBuilder and a separator String

Guava's Joiner type is no longer used in any case, as this not only creates temporary arrays and strings, but is also arguably no more readable. Additionally, the code now combines literals and uses run-on statements where possible, so even the cases that already fell into the categories above are now more readable.

The logic has been moved to a new ToStringGenerator class to reduce the size of the immense CodeGenerator class, as well as make it feasible to test it much more thoroughly.