Add TypeClass
When parameterized with TypeParameterElements, a Type acts like a "prototypical" type of its class, as defined in the JavaDoc for Element.asType. We have been taking advantage of this to combine a TypeElement and its prototypical DeclaredType into a single object, with the declaration and declarationParameters methods returning element-like excerpts, and the remaining methods returning type-like excerpts.

However, this led to subtle bugs when code expecting such a hybrid Type was accidentally fed one created from the prototypical DeclaredType, rather than the TypeElement. Generated code would end up losing bound information—a rare enough edge case that it can slip through the gaps when writing automated tests. As a short-term fix, the methods returning element-like excerpts were modified to throw an exception if the parameter list contained anything other than TypeParameterElements. This ensured any bugs were exhibited in any generic code test, not just tests with bounds. Checking for misuse at runtime is smelly code, so the long-term goal was to push this into the type system.

This change adds an explicit Type subclass, TypeClass, to represent this hybrid object. The two element-style methods have been moved to this class, and the stop-gap state checks removed. Datatype now uses TypeClass to represent the user's type and all generated type; the user's Builder type is still a Type to make clear it is _not_ necessarily prototypical for the Builder.