Track current source scope and emit shorter references to nested types
By keeping track of where we are in the generated code, we can simplify references to sibling types, improving readability.

For instance, we now emit `new Value(this);` instead of `new Foo_Builder.Value(this);`

The core responsibility of tracking block starts and ends, together with detecting class declarations, falls to SourceParser, which (like the SAX framework) emits callbacks to CompilationUnitBuilder to allow it to keep track of the stack of scopes.