Track type body blocks automatically when emitting code
Currently, code generators have to explicitly create new source objects to track when they are generating code inside a type block. They also need to redeclare supertype information that's already been written once to the source. We could make the type block source objects automatically generate the source code, but that requires passing in even more metadata, and modifying all existing code generators to use the new approach. I also struggled to find a general solution that was as readable and flexible as the traditional "append code" approach.

Instead, reparse the generated code to detect {}-delimited blocks, and pull type declarations, including supertype information, back out as well. This lets us keep the code generators clean, and build smarts once-and-only-once in the SourceBuilder framework.

The core responsibility of tracking block starts and ends, together with detecting class declarations, falls to SourceParser, which (like the SAX framework) emits callbacks to CompilationUnitBuilder to allow it to keep track of the stack of scopes.