Use short names for types in current scope
If a type is in scope, use its short name. This can happen for the following reasons:

 * it's part of the current generated type
 * it's a top-level type in the current package
 * it's been pulled in by inheritance

Source tests are not reflecting the full scope of this change, as the scope manager relies on an Elements instance to determine what is in scope, and this is only partially stubbed in these tests.