Allow alternative name for build method
If it would be a compiler error to emit the default build method, whether because the user's Builder declares it with decreased visibility, or changes its return type, instead change the visibility, and if necessary the name, following a predetermined pattern (`_buildImpl`, `_buildImpl2`, etc.).

This allows the user to use FreeBuilder for classes where `build` is not available, albeit at the cost of enhanced support for this type in other FreeBuilder types.