Merge SharedBehaviorTesting into its "subclass"
SharedBehaviorTesting was refactored out of BehaviorTestRunner when ParameterizedBehaviorTestFactory was created on 19 July 2016 (207354b979 - "Support parameterized JUnit tests" - merged as part of PR #192 and released in v1.10.6), but was left behind when BehaviorTestRunner was subsequently removed on 8 Aug 2017 (ea52b751bd - "Set source level in behavior tests" - merged as part of PR #256 and released in v1.14.3).

ParameterizedBehaviorTestRunner was effectively a subclass but (as Java does not allow multiple inheritance) which used composition and constructor parameters to mimic inheritance. Merge the two classes together into a single top-level class.